+++ src/ungod/base/Utility.cpp
@@ -3,7 +3,7 @@
 
 namespace ungod
 {
-    std::mt19937 NumberGenerator::gen(time(0));
+    std::mt19937 NumberGenerator::gen((int)time(0));
 
     int NumberGenerator::getRandBetw(const int a, const int b)
     {
@@ -14,13 +14,13 @@ namespace ungod
     float NumberGenerator::getFloatRandBetw(const float a, const float b)
     {
         std::uniform_real_distribution<> dist(a,b);
-        return dist(gen);
+        return (float)dist(gen);
     }
 
     float NumberGenerator::getNormRand(const float mu, const float rho)
     {
         std::normal_distribution<float> dist(mu, rho);
-        return dist(gen);
+        return (float)dist(gen);
     }
 
     bool NumberGenerator::getRandBool()
